mdadm: Gcc 7 fixes and reautospeced
gcc 7 adds some more diagnostics, and as we compile with diagnostics
becoming errors this causes us to fail to build. The fixes are
increasing a buffer size so gcc knows it will not overflow, adding in
multiple /* Falls through */ comments to document that the lack of
break statement is intentional, and adding gcc pragmas to indicate
that it is deliberate that we don't need the terminating NUL character
in a character array inside a struct.

Redhat have a different approach, they just turn off the warnings in
the Makefile! If their changes get accepted upstream then this patch
should probably be removed. However I think it is better to keep as
many warnings enabled as possible and just disable them locally.

Signed-off-by: Icarus Sparry <icarus.w.sparry@intel.com>